Reorganize routes into named feature folders, add views for routes
